Revert "msm: kgsl: Affine IRQ and worker kthread to the big CPU cluster"
This reverts commit 417bded5a942a2a23ad65b3fe5fd3fff2d0dbf5b.

This is wrong. This causes 3 IRQs to be affined to the big CPU cluster,
not just the primary kgsl_3d0_irq one. As a result, the perf crit API
thinks that the 2 extra IRQs are critical and will balance them despite
them being rarely used (kgsl_hfi_irq and kgsl_gmu_irq).

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: Forenche <prahul2003@gmail.com>